+++ CHANGELOG.md
@@ -11,6 +11,7 @@ No unreleased features
 
 ## Done
 
+* `New-Microservice` requiredRoles are now set when passing the cumulocity.json manifest file to the `-File` parameter
 * Added `New-ServiceUser` and `Get-ServiceUser` to create and get a service user that can be used for automation purposes
 
     ```powershell